# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Gram Pragati Setu API
  version: 1.0.0
  description: Backend API for Gram Pragati Setu web portal
servers:
  - url: http://localhost:5000/api
    description: Development server
paths:
...
  /funds/summary:
    get:
      summary: Fund summary (allocated, utilized, remaining)
      tags: [Funds]
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: village_id
          in: query
          schema:
            type: integer
        - name: project_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Fund summary
  /funds/transactions:
    get:
      summary: List fund transactions with filters
      tags: [Funds]
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: village_id
          in: query
          schema:
            type: integer
        - name: project_id
          in: query
          schema:
            type: integer
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of fund transactions
  /funds/allocate:
    post:
      summary: Allocate funds to a project (Officer only)
      tags: [Funds]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, amount]
              properties:
                project_id:
                  type: integer
                amount:
                  type: number
                tranche:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: Funds allocated
  /funds/release:
    post:
      summary: Mark funds as utilized / released (Officer only)
      tags: [Funds]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, amount]
              properties:
                project_id:
                  type: integer
                amount:
                  type: number
                note:
                  type: string
      responses:
        '200':
          description: Funds released
  /funds/{id}:
    patch:
      summary: Edit a fund transaction (Officer only)
      tags: [Funds]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                note:
                  type: string
      responses:
        '200':
          description: Transaction updated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT